+++ packages/orm/src/client/crud/dialects/sqlite.ts
@@ -10,11 +10,11 @@ import {
     type SqlBool,
 } from 'kysely';
 import { match } from 'ts-pattern';
+import { AnyNullClass, DbNullClass, JsonNullClass } from '../../../common-types';
 import type { BuiltinType, FieldDef, GetModels, SchemaDef } from '../../../schema';
 import { DELEGATE_JOINED_FIELD_PREFIX } from '../../constants';
 import type { FindArgs } from '../../crud-types';
 import { createInternalError } from '../../errors';
-import { AnyNullClass, DbNullClass, JsonNullClass } from '../../null-values';
 import {
     getDelegateDescendantModels,
     getManyToManyRelation,
