+++ packages/orm/src/index.ts
@@ -1,4 +1,4 @@
 export * from './client';
+export * from './common-types';
 export * as KyselyUtils from './utils/kysely-utils';
 export * as SchemaUtils from './utils/schema-utils';
-export type { JsonArray, JsonObject, JsonValue } from './utils/type-utils';
